--- src/components/App/App.js
@@ -1,9 +1,0 @@
-import React from "react";
-import Main from "../mainPage/Main/Main";
-
-export default function App() {
-  return (
-    <Main />
-  );
-}
-
