+++ src/components/App/app.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import Main from "../main-page/main/main";
+
+export default function App() {
+  return <Main />;
+}
