+++ src/components/main-page/hotel-card/hotel-card.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import HotelCard from './HotelCard';
+import HotelCard from './hotel-card';
 
 
 Enzyme.configure({adapter: new Adapter()});
