+++ src/components/main-page/hotel-card/hotel-card.jsx
