--- src/components/mainPage/City/City.js
@@ -1,82 +1,0 @@
-import React from "react";
-import HotelCard from "../HotelCard/HotelCard";
-
-const hotels = [{
-  isPremium: true,
-  cost: 120,
-  name: `Beautiful & luxurious apartment at great location`,
-  rating: 55,
-  type: `Apartment`,
-}, {
-  isPremium: false,
-  cost: 132,
-  name: `Wood and stone place`,
-  rating: 80,
-  type: `Apartment`,
-}, {
-  isPremium: true,
-  cost: 120,
-  name: `Canal View Prinsengracht`,
-  rating: 55,
-  type: `Apartment`,
-}, {
-  isPremium: true,
-  cost: 80,
-  name: `Nice, cozy, warm big bed apartment`,
-  rating: 100,
-  type: `Apartment`,
-}, {
-  isPremium: true,
-  cost: 120,
-  name: `Wood and stone place`,
-  rating: 88,
-  type: `Private room`,
-},
-];
-
-function handleClick() {
-
-
-}
-
-export default function City() {
-  return (
-    <div className="cities">
-      <div className="cities__places-container container">
-        <section className="cities__places places">
-          <h2 className="visually-hidden">Places</h2>
-          <b className="places__found">312 places to stay in Amsterdam</b>
-          <form className="places__sorting" action="#" method="get">
-            <span className="places__sorting-caption">Sort by</span>
-            <span className="places__sorting-type" tabIndex="0">
-              Popular<svg className="places__sorting-arrow" width="7" height="4">
-                <use xlinkHref="#icon-arrow-select"></use>
-              </svg>
-            </span>
-            <ul className="places__options places__options--custom places__options--opened">
-              <li className="places__option places__option--active" tabIndex="0">Popular</li>
-              <li className="places__option" tabIndex="0">Price: low to high</li>
-              <li className="places__option" tabIndex="0">Price: high to low</li>
-              <li className="places__option" tabIndex="0">Top rated first</li>
-            </ul>
-          </form>
-          <div className="cities__places-list places__list tabs__content">
-            {hotels.map((it, i) =>
-              <HotelCard
-                key={name + i}
-                name={it.name}
-                isPremium={it.isPremium}
-                cost={it.cost}
-                rating={it.rating}
-                type={it.type}
-                onClick={handleClick}
-              />)}
-          </div>
-        </section>
-        <div className="cities__right-section">
-          <section className="cities__map map"></section>
-        </div>
-      </div>
-    </div >
-  );
-}
