+++ src/components/tabs-panel/locations-city/locations-city.e2e.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import LocationsCity from './LocationsCity';
+import LocationsCity from './locations-city';
 
 
 Enzyme.configure({adapter: new Adapter()});
