+++ src/components/tabs-panel/locations-city/locations-city.jsx
