+++ src/components/tabs-panel/tabs/tabs.test.js
@@ -1,0 +1,11 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import Tabs from './tabs';
+
+it(`Tabs correctly renders after relaunch`, () => {
+  const tree = renderer
+    .create(<Tabs />)
+
+    .toJSON();
+  expect(tree).toMatchSnapshot();
+});
