--- src/index.js
@@ -1,14 +1,0 @@
-import React from "react";
-import ReactDOM from 'react-dom';
-import App from './components/App/App';
-
-
-const init = () => {
-  ReactDOM.render(
-      <App />,
-      document.querySelector(`#root`)
-
-  );
-};
-
-init();
